Replace usage of onViewCreated() with onViewStateRestored()
It is not safe to use viewLifecycleOwner or its lifecycleScope,
since it is possible for Fragment to call its onViewCreated()
and then almost immediately onDestroyView() without its view
actually leaving INITIALIZED state. This results in not entering
DESTROYED state and viewLifecycleOwner.lifecycleScope coroutines
not being cancelled.

However, if onViewStateRestored() was called, we know for sure that
Fragment's view will enter CREATED state right after.