Add workaround for Let's Encrypt certificates on Android < 7.1.1 (#60)
Despite the fact that DST Root CA X3 cartificate (which
is root certificate for Let's Encrypt) is expired, it is supposed
to still work on Android since Android doesn't check expiration date
of root certificates. However, this is true only for Android's Java network stack
and we use Qt + OpenSSL for networking instead. OpenSSL does check for root
certificates expiration, with no way to disable that (besides patching, and I don't want to do that).
To fix compatibility with Let's Encrypt certificates, bundle ISRG Root X1
(new Let's Encrypt root certificate which currently actually intermediate and is signed by DST Root CA X3)
with libtremotesf and add it to the list of root certificates.